[testsuite][pangu][mptl] Renamed pangu directory pangu_lib
