--- testsuite/pangu/__init__.py
@@ -1,8 +1,0 @@
-import os
-import sys
-
-
-path = os.path.dirname(__file__)
-
-if path not in sys.path:
-    sys.path.append(path)
