+++ testsuite/pangu_lib/fixtures/pfn_1-statefulset.yaml
