+++ testsuite/pangu_lib/fixtures/stateful_vfn_1.yaml
