+++ testsuite/pangu_lib/node_commands/stop_node.py
