+++ testsuite/pangu_lib/template_testnet_files/pangu_node_config.yaml
