+++ testsuite/pangu_lib/template_testnet_files/validator.yaml
