+++ testsuite/pangu_lib/testnet_commands/delete_testnet.py
